fix paginas ajustadas

+++ src/postagem/services/postagem.service.ts
@@ -1,3 +1,4 @@
+import { TemaService } from './../../tema/services/tema.service';
 import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { DeleteResult, ILike, Repository } from 'typeorm';
@@ -8,13 +9,14 @@ export class PostagemService {
   constructor(
     @InjectRepository(Postagem)
     private postagemRepository: Repository<Postagem>,
+    private temaService: TemaService,
   ) {}
+
   async findAll(): Promise<Postagem[]> {
     return await this.postagemRepository.find({
-      relations:{
-                tema: true,
-                usuario:true
-      }
+      relations: {
+        tema: true,
+      },
     });
   }
 
@@ -23,10 +25,9 @@ export class PostagemService {
       where: {
         id,
       },
-      relations:{
-                tema: true,
-                usuario:true
-      }
+      relations: {
+        tema: true,
+      },
     });
 
     if (!postagem)
@@ -38,27 +39,51 @@ export class PostagemService {
   async findAllByTitulo(titulo: string): Promise<Postagem[]> {
     return await this.postagemRepository.find({
       where: {
-        titulo: ILike(`%${titulo}%`),  //a % significa que tanto faz o que tenha antes do que quero pesquisar e tanto faz o que esta depois, o que eu quero é aquela letra/palavra que inseri, aqui o ILike ele não se preocupa se a pessoa digitar com letra maiúscula ou minúscula, já o Like se preocupa.
+        titulo: ILike(`%${titulo}%`), //a % significa que tanto faz o que tenha antes do que quero pesquisar e tanto faz o que esta depois, o que eu quero é aquela letra/palavra que inseri, aqui o ILike ele não se preocupa se a pessoa digitar com letra maiúscula ou minúscula, já o Like se preocupa.
       },
-      relations:{
-                tema: true,
-                usuario:true
-      }
+      relations: {
+        tema: true,
+      },
     });
   }
 
-    async create(postagem: Postagem): Promise<Postagem> {
-        return await this.postagemRepository.save(postagem);
+  async create(postagem: Postagem): Promise<Postagem> {
+    if (postagem.tema) {
+      let tema = await this.temaService.findById(postagem.tema.id);
+
+      if (!tema)
+        throw new HttpException('Tema não encontrado!', HttpStatus.NOT_FOUND);
+
+      return await this.postagemRepository.save(postagem);
     }
 
-    async update(postagem: Postagem): Promise<Postagem>{
-        await this.findById(postagem.id)
-        return await this.postagemRepository.save(postagem);
+    return await this.postagemRepository.save(postagem);
+  }
+
+  async update(postagem: Postagem): Promise<Postagem> {
+    let buscaPostagem: Postagem = await this.findById(postagem.id);
+
+    if (!buscaPostagem || !postagem.id)
+      throw new HttpException('Postagem não encontrada!', HttpStatus.NOT_FOUND);
+
+    if (postagem.tema) {
+      let tema = await this.temaService.findById(postagem.tema.id);
+
+      if (!tema)
+        throw new HttpException('Tema não encontrado!', HttpStatus.NOT_FOUND);
+
+      return await this.postagemRepository.save(postagem);
     }
 
-    async delete(id: number): Promise<DeleteResult> {
-        await this.findById(id);
-        return await this.postagemRepository.delete(id);
-    }
+    return await this.postagemRepository.save(postagem);
+  }
 
+  async delete(id: number): Promise<DeleteResult> {
+    let buscaPostagem = await this.findById(id);
+
+    if (!buscaPostagem)
+      throw new HttpException('Postagem não encontrada!', HttpStatus.NOT_FOUND);
+
+    return await this.postagemRepository.delete(id);
+  }
 }
